Change the license from GPL to MIT License
*** Why can we switch the license?

1. Our only two pull requests are licensed under the MIT License. (#2 and #3, I specifically asked the committer to license them under the MIT license to me.)

2. The rest of the code is written by me so I can choose a new license for them.

*** What are the benefits of the license changing?

1. This allows us to accept pull requests without making the BTCTools GUI open source under the GPL.

2. You can also use these codes for your commercial projects without having to open source them.

*** Why doesn't the copyleft feature of GPL prevent us from license switching?

In fact, any GPL project can switch to another license in the future - if the proposal is agreed to by all authors.

The GPL is a contract between an author and other authors for a joint work. But this contract is not exclusive. As long as all authors are willing, they are entirely free to enter into another contract for their joint work. This is how some "dual-licensing" commercial projects work. This is also my pattern of accepting #2 and #3 (I got a second license from the author).

So, given that I'm the author of the code for this project except #2 and #3, and I hold the MIT license for the code #2 and #3 (so I can sublicense them), I think I have the right to choose for all the code a new license.

Of course, this doesn't end the GPL license for existing code. If you are a free software enthusiast, you can continue to receive and sublicense the code under the GPL.

But new code -- especially from outside developers -- will be received under the MIT license. This will allow us to have a simpler workflow without having to seek dual licensing in order to publish a closed source application. Of course, this also means that we relinquish control of that code. Anyone can use it for commercial development without exposing the source code to us.